+++ resources/views/components/text/heading.blade.php
@@ -9,16 +9,8 @@ $element = $asHeading ? 'h'.$headingLevel : 'div';
 
     'font-semibold' => ! BladeComponents::containsFontWeightClass($attributes->get('class')),
 
-    'text-lg' => ! BladeComponents::containsFontSizeClass($attributes->get('class'))
-        && $size === 'lg'
-        && $size !== 'none',
-    'text-base' => ! BladeComponents::containsFontSizeClass($attributes->get('class'))
-        && is_null($size)
-        && $size !== 'none',
-    'text-sm' => ! BladeComponents::containsFontSizeClass($attributes->get('class'))
-        && $size === 'sm'
-        && $size !== 'none',
-    'text-xs' => ! BladeComponents::containsFontSizeClass($attributes->get('class'))
-        && $size === 'xs'
-        && $size !== 'none',
+    'text-lg' => ! BladeComponents::containsFontSizeClass($attributes->get('class')) && $size === 'lg',
+    'text-base' => ! BladeComponents::containsFontSizeClass($attributes->get('class')) && $size === 'default',
+    'text-sm' => ! BladeComponents::containsFontSizeClass($attributes->get('class')) && $size === 'sm',
+    'text-xs' => ! BladeComponents::containsFontSizeClass($attributes->get('class')) && $size === 'xs',
 ]) }}>{{ $slot }}</{{ $element }}>